Create minimax_ttt.py
The minimax algorithm works for every move except the first move. The first move uses a depth of 8, and because of the recursion, takes too long to find the best move, therefore I lessened the depth on the first move. However, because of this, the first move is not necessarily the best move. I may change this later with alpha beta pruning or by making my program more efficient. For the rest of the moves, the program works fine.